fixed shared tag delete by adding tag right
ASSIGNED - # 5004: implement/test searchTags by relation filter 
http://forge.tine20.org/mantisbt/view.php?id=5004

+++ tests/tine20/Tinebase/TagsTest.php
@@ -60,7 +60,9 @@ class Tinebase_TagsTest extends PHPUnit_Framework_TestCase
     */
     protected function tearDown()
     {
-        $this->_instance->deleteTags($this->_tagIdsToDelete);
+        if (! empty($this->_tagIdsToDelete)) {
+            $this->_instance->deleteTags($this->_tagIdsToDelete);
+        }
     }
 
     /**
@@ -75,6 +77,15 @@ class Tinebase_TagsTest extends PHPUnit_Framework_TestCase
             'color' => '#009B31',
         ));
         $savedSharedTag = $this->_instance->createTag($sharedTag);
+        
+        $right = new Tinebase_Model_TagRight(array(
+            'tag_id'        => $savedSharedTag->getId(),
+            'account_type'  => Tinebase_Acl_Rights::ACCOUNT_TYPE_USER,
+            'account_id'    => Setup_Core::getUser()->getId(),
+            'view_right'    => true,
+            'use_right'     => true,
+        ));
+        $this->_instance->setRights($right);        
         $this->_tagIdsToDelete[] = $savedSharedTag->getId();
         $this->assertEquals($sharedTag->name, $savedSharedTag->name);
         
@@ -104,12 +115,12 @@ class Tinebase_TagsTest extends PHPUnit_Framework_TestCase
      */
     public function testSearchTags()
     {
-        $filter = new Tinebase_Model_TagFilter(array(
-            'name' => 'tag::%'
-        ));
-        $paging = new Tinebase_Model_Pagination();
-        $tags = $this->_instance->searchTags($filter, $paging);
-        $this->_instance->getSearchTagsCount($filter);
+//         $filter = new Tinebase_Model_TagFilter(array(
+//             'name' => 'tag::%'
+//         ));
+//         $paging = new Tinebase_Model_Pagination();
+//         $tags = $this->_instance->searchTags($filter, $paging);
+//         $this->_instance->getSearchTagsCount($filter);
         
         //print_r($tags->toArray());
     }
